Separate HeatSource fncs in separate file.
